Fixed uint64 to int conversion issue by replacing all height types to uint64. Renamed all "min" and "max" names to avoid collitions with the built-in functions.

+++ pkg/analysis/criteria/height_test.go
@@ -14,7 +14,7 @@ import (
 )
 
 type heightInfo struct {
-	Height int
+	Height uint64
 	Nodes  entities.Nodes
 }
